+++ NEWS.md
@@ -1,7 +1,18 @@
 
 # cqcr 0.1.2.9000
 
-New `cqc_location_details` function.
+* New `cqc_location_details` function.
+
+## Internal changes
+
+* Reduced json page sizes on calls
+
+* Code-reorganising to streamline maintenance
+
+## Non-breaking changes
+
+* Package now displays a message if cqc_partner_code is not set. Message will 
+be displayed even if `verbose=FALSE`.
 
 
 # cqcr 0.1.2